chore(api): separate workspace and workspace record to two different files
