+++ leptonai/api/v1/workspace.py
@@ -35,112 +35,7 @@ from .deployment import DeploymentAPI
 from .job import JobAPI
 from .secret import SecretAPI
 
-
-class WorkspaceRecord(object):
-    """
-    Internal class to manage the local
-    """
-
-    _singleton_dict: Dict[str, Any] = {"workspaces": {}, "current_workspace": None}
-    # global lock for reading and writing the workspace info file
-    _rw_lock = Lock()
-    WORKSPACE_FILE = CACHE_DIR / "workspace_info.yaml"
-
-    def __init__(self):
-        raise RuntimeError("WorkspaceInfoLocalRecord should not be instantiated.")
-
-    @classmethod
-    def load_workspace_info(cls):
-        if cls.WORKSPACE_FILE.exists():
-            with cls._rw_lock:
-                with open(cls.WORKSPACE_FILE) as f:
-                    cls._singleton_dict = yaml.safe_load(f)
-
-    @classmethod
-    def reload(cls):
-        cls.load_workspace_info()
-
-    @classmethod
-    def _save_to_file(cls):
-        create_cached_dir_if_needed()
-        with cls._rw_lock:
-            with open(cls.WORKSPACE_FILE, "w") as f:
-                yaml.safe_dump(cls._singleton_dict, f)
-
-    @classmethod
-    def set_and_save(
-        cls,
-        workspace_id: str,
-        auth_token: Optional[str] = None,
-        url: Optional[str] = None,
-    ):
-        """
-        Saves a workspace by adding it to the workspace info file.
-        """
-        try:
-            display_name = _get_workspace_display_name(workspace_id)
-        except RuntimeError:
-            display_name = ""
-        cls._singleton_dict["workspaces"][workspace_id] = {}
-        if url is None:
-            url = _get_full_workspace_api_url(workspace_id)
-        cls._singleton_dict["workspaces"][workspace_id]["url"] = url
-        cls._singleton_dict["workspaces"][workspace_id]["display_name"] = display_name
-        cls._singleton_dict["workspaces"][workspace_id]["auth_token"] = auth_token
-        cls.set_current(workspace_id)
-        cls._save_to_file()
-
-    @classmethod
-    def has(cls, workspace_id: str):
-        return workspace_id in cls._singleton_dict["workspaces"]
-
-    @classmethod
-    def _current_workspace_id(cls) -> Union[str, None]:
-        return cls._singleton_dict["current_workspace"]
-
-    @classmethod
-    def get(cls, workspace_id: str):
-        try:
-            ws = cls._singleton_dict["workspaces"][workspace_id]
-        except KeyError:
-            raise
-        return Workspace(workspace_id, ws["auth_token"], ws["url"])
-
-    @classmethod
-    def current(cls):
-        name = cls._current_workspace_id()
-        if name is None:
-            raise RuntimeError("You have not set the current workspace yet.")
-        else:
-            return cls.get(name)
-
-    @classmethod
-    def set_current(cls, workspace_id: Optional[str] = None):
-        """
-        Sets the current workspace to the given workspace_id, or None if no workspace_id is given.
-        """
-        if workspace_id and workspace_id not in cls._singleton_dict["workspaces"]:
-            raise ValueError(f"Workspace {workspace_id} does not exist.")
-        cls._save_to_file()
-
-    @classmethod
-    def remove(cls, workspace_id: str):
-        """
-        Removes the workspace with the given workspace_id.
-        """
-        cls._singleton_dict["workspaces"].pop(workspace_id)
-        if cls._singleton_dict["current_workspace"] == workspace_id:
-            cls._singleton_dict["current_workspace"] = None
-        cls._save_to_file()
-
-
-# When importing, read the content of the workspace info file as initialization.
-WorkspaceRecord.load_workspace_info()
-
-
-_semver_pattern = re.compile(
-    r"^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$"  # noqa: W605
-)
+from .workspace_record import WorkspaceRecord
 
 
 class Workspace(object):
@@ -257,6 +152,9 @@ class Workspace(object):
     T = TypeVar("T", bound=BaseModel)
 
     def ensure_type(self, response, EnsuredType: Type[T]) -> T:
+        """
+        Utility function to ensure that the response is of the given type.
+        """
         if not response.ok:
             raise RuntimeError(
                 f"API call failed with status code {response.status_code}. Details:"
@@ -277,6 +175,9 @@ class Workspace(object):
             )
 
     def ensure_list(self, response, EnsuredType: Type[T]) -> List[T]:
+        """
+        Utility function to ensure that the response is a list of the given type.
+        """
         if not response.ok:
             raise RuntimeError(
                 f"API call failed with status code {response.status_code}. Details:"
@@ -296,7 +197,10 @@ class Workspace(object):
                 f" details:\n{e}\n*** end of debug info ***"
             )
 
-    def ensure_ok(self, response):
+    def ensure_ok(self, response) -> bool:
+        """
+        Utility function to ensure that the response is ok.
+        """
         if not response.ok:
             raise RuntimeError(
                 f"API call failed with status code {response.status_code}. Details:"
@@ -304,7 +208,10 @@ class Workspace(object):
             )
         return True
 
-    def ensure_json(self, response):
+    def ensure_json(self, response) -> Union[Dict, List, str]:
+        """
+        Utility function to ensure that the output is a json object (including dict, list, etc.)
+        """
         if not response.ok:
             raise RuntimeError(
                 f"API call failed with status code {response.status_code}. Details:"
@@ -337,6 +244,10 @@ class Workspace(object):
         this is a dev workspace, returns None.
         """
         info = self.info()
+        _semver_pattern = re.compile(
+            r"^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$"  # noqa: W605
+        )
+
         match = _semver_pattern.match(info.git_commit)
         return (
             (int(match.group(1)), int(match.group(2)), int(match.group(3)))
@@ -346,6 +257,9 @@ class Workspace(object):
 
 
 def current() -> Workspace:
+    """
+    Returns the current workspace.
+    """
     return WorkspaceRecord.current()
 
 
@@ -355,7 +269,7 @@ def login(
     url: Optional[str] = None,
 ) -> None:
     """
-    Logs in to a workspace.
+    Logs in to a workspace and saves the workspace information.
     """
     if workspace_id is None:
         if WorkspaceRecord._current_workspace_id is None:
